Create conversion model

+++ spec/factories/currencies.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :currency do
     iso_code { Faker::Currency.code }
-    rate { rand(1.1..100).to_f }
+    rate { rand(1.1..100).to_d }
     is_base { Faker::Boolean.boolean }
   end
 end
